backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "asyncio>=4.0.0",
    "fastapi>=0.127.0",
    "python-multipart>=0.0.9",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "asyncpg>=0.31.0",
    "greenlet>=3.3.0",
    "pyjwt>=2.10.1",
    "pwdlib[argon2]>=0.3.0",
    "fernet>=1.0.1",
    "requests>=2.32.5",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "scikit-learn>=1.8.0",
    "statsmodels>=0.14.6",
    "dpkt>=1.9.8",
    "core",
    "alembic>=1.17.2",
    "cryptography>=46.0.3",
    "minio>=7.2.20",
    "sniffio>=1.3.1",
]

[tool.uv.sources]
core = { path = "../core", editable = true }
